+++ tests/prefetch/test_prefetch_m2m.py
@@ -11,14 +11,23 @@ database = DatabaseTestClient(DATABASE_URL)
 models = edgy.Registry(database=edgy.Database(database, force_rollback=True))
 
 
-class User(edgy.StrictModel):
-    name = edgy.CharField(max_length=100)
-
+class IntrospectingModel(edgy.StrictModel):
     class Meta:
         registry = models
+        abstract = True
+
+    @classmethod
+    async def from_sqla_row(cls, **kwargs) -> edgy.Model:
+        returnobj = await super().from_sqla_row(**kwargs)
+        object.__setattr__(returnobj, "introspected_prefetches", kwargs.get("prefetch_related"))
+        return returnobj
 
 
-class Role(edgy.StrictModel):
+class User(IntrospectingModel):
+    name = edgy.CharField(max_length=100)
+
+
+class Role(IntrospectingModel):
     name = edgy.CharField(max_length=100)
     users = edgy.ManyToMany("User", related_name="roles")
 
@@ -26,7 +35,7 @@ class Role(edgy.StrictModel):
         registry = models
 
 
-class SpaceGroup(edgy.StrictModel):
+class SpaceGroup(IntrospectingModel):
     role = edgy.ForeignKey(Role, related_name="groups")
     name = edgy.CharField(max_length=100)
 
@@ -34,7 +43,7 @@ class SpaceGroup(edgy.StrictModel):
         registry = models
 
 
-class Space(edgy.StrictModel):
+class Space(IntrospectingModel):
     groups = edgy.ManyToManyField(SpaceGroup, related_name="spaces")
     name = edgy.CharField(max_length=100)
 
@@ -69,11 +78,12 @@ async def test_prefetch_m2m_directly():
 
     assert len(total) == 2
 
-    space_query = await Space.query.prefetch_related(
-        Prefetch(related_name="groups", to_attr="to_groups")
-    ).get(id=space.id)
+    prefetches = [Prefetch(related_name="groups", to_attr="to_groups")]
 
+    space_query = await Space.query.prefetch_related(*prefetches).get(id=space.id)
     assert len(space_query.to_groups) == 2
+    assert all(not prefetch._baked and not prefetch._baked_results for prefetch in prefetches)
+    assert all(prefetch._baked_results for prefetch in space_query.introspected_prefetches)
 
 
 async def test_prefetch_m2m_directly_mixed():
@@ -90,16 +100,46 @@ async def test_prefetch_m2m_directly_mixed():
     await role.users.add_many(user, user2)
 
     assert len(total) == 2
-
-    space_query = await Space.query.prefetch_related(
+    prefetches = [
         Prefetch(
             related_name="groups__role__users",
             to_attr="to_users",
             queryset=User.query.all().distinct("id"),
         )
-    ).get(id=space.id)
+    ]
+
+    space_query = await Space.query.prefetch_related(*prefetches).get(id=space.id)
 
     assert len(space_query.to_users) == 2
+    assert all(not prefetch._baked and not prefetch._baked_results for prefetch in prefetches)
+    assert all(prefetch._baked_results for prefetch in space_query.introspected_prefetches)
+
+
+async def test_prefetch_m2m_directly_mixed_none():
+    role = await Role.query.create(name="Admin")
+    group1 = await SpaceGroup.query.create(name="Group 1", role=role)
+    group2 = await SpaceGroup.query.create(name="Group 2", role=role)
+
+    space = await Space.query.create(name="Space 1")
+    total = await space.groups.add_many(group1, group2)
+
+    await User.query.create(name="Edgy")
+    await User.query.create(name="Ravyn")
+
+    assert len(total) == 2
+    prefetches = [
+        Prefetch(
+            related_name="groups__role__users",
+            to_attr="to_users",
+            queryset=User.query.all().distinct("id"),
+        )
+    ]
+
+    space_query = await Space.query.prefetch_related(*prefetches).get(id=space.id)
+
+    assert len(space_query.to_users) == 0
+    assert all(not prefetch._baked and not prefetch._baked_results for prefetch in prefetches)
+    assert all(prefetch._baked_results for prefetch in space_query.introspected_prefetches)
 
 
 async def test_prefetch_m2m_reverse_prefetch():
@@ -116,13 +156,40 @@ async def test_prefetch_m2m_reverse_prefetch():
     await role.users.add_many(user, user2)
 
     assert len(total) == 2
-
-    user_query = await User.query.prefetch_related(
+    prefetches = [
         Prefetch(
             related_name="roles__groups__spaces",
             to_attr="to_spaces",
             queryset=Space.query.all().distinct("id"),
         )
-    ).get(id=user.id)
+    ]
+
+    user_query = await User.query.prefetch_related(*prefetches).get(id=user.id)
 
     assert len(user_query.to_spaces) == 1
+    assert all(not prefetch._baked and not prefetch._baked_results for prefetch in prefetches)
+    assert all(prefetch._baked_results for prefetch in user_query.introspected_prefetches)
+
+
+async def test_prefetch_m2m_reverse_prefetch_none():
+    role = await Role.query.create(name="Admin")
+    group1 = await SpaceGroup.query.create(name="Group 1", role=role)
+    group2 = await SpaceGroup.query.create(name="Group 2", role=role)
+
+    space = await Space.query.create(name="Space 1")
+    total = await space.groups.add_many(group1, group2)
+
+    user = await User.query.create(name="Edgy")
+    await User.query.create(name="Ravyn")
+
+    assert len(total) == 2
+
+    user_query = await User.query.prefetch_related(
+        Prefetch(
+            related_name="roles__groups__spaces",
+            to_attr="to_spaces",
+            queryset=Space.query.filter(id=-100),
+        )
+    ).get(id=user.id)
+
+    assert len(user_query.to_spaces) == 0